requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.0
drf_yasg
djangorestframework-simplejwt